fix #2884 make sure that error objects are deserialized as errors not as aliases, earlier in the 5.x series this was always a string
